Refactor arg conversion and ground into separate files.
