+++ query-algebrizer/src/clauses/mod.rs
@@ -61,12 +61,15 @@ use types::{
     ValueTypeSet,
 };
 
+mod convert;              // Converting args to values.
 mod inputs;
 mod or;
 mod not;
 mod pattern;
 mod predicate;
 mod resolve;
+
+mod ground;
 mod where_fn;
 
 use validate::{
